Make display more generic
Make the draw_screen callback more generic, so that emulators have more
freedom in how they implement the screen. This should also make it
easier to change away from minifb later on if needed.

Adjust chip8 to store the screen as a bitvector, which is converted to
white pixels when drawing.